#2963: Constant evaluation for external function calls. Reverting r6048 since some newly introduced tests doesnt work on linux.
git-svn-id: https://svn.jmodelica.org@6055 7ae51c65-13d5-4921-9d80-6511ddd1419c

+++ trunk/Python/src/tests_jmodelica/files/Modelica/Resources/src/arrayFunctions.c
@@ -25,7 +25,7 @@ void transposeMatrix(double* a, size_t a_rows, size_t a_cols, double* b, size_t 
     }
 }
 
-void extFunc1(double m, double* a, size_t a1, size_t a2, size_t a3, int* b, size_t b1, size_t b2, size_t b3, int* c, size_t c1, size_t c2, double* sum, 
+void extFunc1(double m, double* a, size_t a1, size_t a2, size_t a3, int* b, size_t b1, size_t b2, size_t b3, double* c, size_t c1, size_t c2, double* sum, 
     double* o, size_t o1, size_t o2, size_t o3, double* step, size_t step1)
 {
     size_t i1,i2,i3,t1,t2,i;
@@ -35,18 +35,18 @@ void extFunc1(double m, double* a, size_t a1, size_t a2, size_t a3, int* b, size
         for (i2 = 0; i2 < a2; i2++) {
             t1 = i1*a2+i2;
             for (i3 = 0; i3 < a3; i3++) {
-                t2 = t1*a3 + i3;
+				t2 = t1*a3 + i3;
                 o[t2] = m*a[t2] / b[t2];
                 if (c[t1])
                     *sum += o[t2];
                 step[i] = *sum;
-                i = i + 1;
+				i = i + 1;
             }
         }
     }
 }
 
-void copyBoolArray(int* bin, size_t bin_size, int* bout, size_t bout_size)
+void copyBoolArray(double* bin, size_t bin_size, double* bout, size_t bout_size)
 {
     size_t i1;
     for (i1 = 0; i1 < bin_size; i1++) {
